--- examples/app-crm/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from "./graphql";
