+++ examples/app-crm/src/routes/dashboard/components/queries.ts
@@ -41,22 +41,3 @@ export const DASHBOARD_TASKS_CHART_QUERY = gql`
         }
     }
 `;
-
-export const DASHBOARD_TOTAL_REVENUE_QUERY = gql`
-    query DashboardTotalRevenue(
-        $filter: DealStageFilter!
-        $sorting: [DealStageSort!]
-        $paging: OffsetPaging!
-    ) {
-        dealStages(filter: $filter, sorting: $sorting, paging: $paging) {
-            nodes {
-                title
-                dealsAggregate {
-                    sum {
-                        value
-                    }
-                }
-            }
-        }
-    }
-`;
